Rename src2 to src.
