--- src/commentsPopup/commentPopup.js
@@ -1,1 +1,0 @@
-import './commentPopupStyles.css';
