+++ src/counterFunctions/commentsCounter.js
