+++ src/index.html
@@ -5,9 +5,10 @@
   <meta charset="UTF-8" />
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <script src="index.js" defer></script>
+  <script src="index.js" type="module" defer></script>
   <!-- <script src="https://cdn.tailwindcss.com"></script> -->
   <script src="https://kit.fontawesome.com/9b54de88d2.js" crossorigin="anonymous"></script>
+  <link rel="stylesheet" href="style.css">
   <title>TV Zone</title>
 </head>
 
@@ -58,6 +59,32 @@
       </li>
     </ul>
   </footer>
+  <!-- <div class="popup">
+    <button class="close-btn"></button>
+    <section class="show-summary">
+      <div class="shows-poster">
+        <img src="" alt="show poster">
+      </div>
+      <div class="show-info">
+        <h2  class="title">title</h2>
+        <ul class="genres"></ul>
+        <div>Rating: <span class="rating">rating</span></div>
+        <p class="summary">add summary here</p>
+      </div>
+    </section>
+    <section class="add-comment-section">
+      <h2>Add comment: </h2>
+      <form action="">
+        <input type="text">
+        <textarea name="" id="" cols="30" rows="10"></textarea>
+        <button></button>
+      </form>
+    </section>
+    <section class="comments-section">
+      <h2>Comments: </h2>
+      <ul class="comments-list"></ul>
+    </section>
+  </div> -->
 </body>
 
 </html>
