--- src/reservationPopup/modules/helperFunctions/listener.js
@@ -1,7 +1,0 @@
-const listener = (btn, event, action) => {
-  btn.addEventListener(event, () => {
-    action();
-  });
-};
-
-export default listener;
