--- src/reservationPopup/reservationPopupStyles.css
@@ -1,213 +1,0 @@
-.popup {
-  position: fixed;
-  z-index: 5;
-  padding-top: 27px;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  overflow: auto;
-  background-color: rgba(0, 0, 0, 0.4);
-}
-
-.popup-container {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  align-content: center;
-  background-color: #8d2323 !important;
-  width: 550px;
-  position: relative;
-  margin: auto;
-  border: 1px solid #888;
-  border-radius: 8px;
-  animation-name: animitop;
-  animation-duration: 0.9s;
-  max-height: calc(100vh - 60px);
-  overflow-y: unset;
-  gap: 10px;
-  padding-bottom: 0;
-  z-index: 1;
-}
-
-.bg-img {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  opacity: 0.1;
-  z-index: -2;
-}
-
-.img {
-  height: 280px;
-  width: 180px;
-  border-radius: 20px;
-  z-index: 99;
-  box-shadow:
-    rgba(0, 0, 0, 0.09) 0 3px 2px,
-    rgba(0, 0, 0, 0.09) 0 6px 3px,
-    rgba(0, 0, 0, 0.09) 0 12px 6px,
-    rgba(0, 0, 0, 0.09) 0 16px 8px,
-    rgba(0, 0, 0, 0.09) 0 32px 16px;
-}
-
-.reservation-form {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 12px;
-}
-
-.close-icon {
-  align-self: flex-end;
-  cursor: pointer;
-  padding: 11px;
-  z-index: 999;
-}
-
-.details-grid {
-  display: grid;
-  grid-template-columns: auto auto auto;
-  list-style: none;
-  column-gap: 12px;
-  padding: 0;
-}
-
-.details-section {
-  grid-column: 1/4;
-  text-align: start;
-  font-size: 14px;
-  font-weight: 700;
-  padding: 3px 0;
-  z-index: 999;
-}
-
-.genre-item {
-  background-color: #d9d9b2;
-  padding: 0 6px;
-  border-radius: 15px;
-  font-size: 12px;
-  font-weight: 500;
-  z-index: 999;
-}
-
-.img-label {
-  font-size: 25px;
-  font-weight: 500;
-  margin: 0;
-
-  /* border: 1px solid #d9d9b2; */
-  padding: 6px;
-  border-radius: 30px;
-  box-shadow: rgba(178, 176, 112, 0.16) 0 3px 6px, rgba(0, 0, 0, 0.23) 0 3px 6px;
-}
-
-.show {
-  display: flex;
-  flex-direction: row;
-  gap: 24px;
-  width: 94%;
-}
-
-.show-details {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.border {
-  grid-column: 1/4;
-  background-color: rgb(160, 64, 64);
-  height: 1px;
-  width: 45px;
-  margin: 0;
-}
-
-.reserve {
-  background-color: #d9d9b2;
-  border: none;
-  width: 90px;
-  height: 30px;
-  border-radius: 15px;
-  color: white;
-  cursor: pointer;
-  transition: 0.5s ease;
-}
-
-.reserve:hover {
-  background: transparent;
-  border: 2px solid #d9d9b2;
-  box-shadow: 0 0 10px 0 #d9d9b2 inset, 0 0 10px 4px #d9d9b2;
-  color: #bdbd93;
-}
-
-input {
-  min-width: 190px;
-  width: 190px;
-  height: 30px;
-  align-self: start;
-  box-shadow: rgba(0, 0, 0, 0.19) 0 10px 20px, rgba(0, 0, 0, 0.23) 0 6px 6px;
-  border: none;
-  padding: 0 10px;
-  border-radius: 15px;
-}
-
-.reservation-item {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  justify-content: space-around;
-  padding-top: 15px;
-}
-
-.reservation-name {
-  font-size: 15px;
-  font-weight: 600;
-}
-
-.reservation-dates {
-  font-size: 12px;
-}
-
-.reservation {
-  display: flex;
-  flex-direction: row;
-  gap: 20px;
-  justify-content: space-around;
-  justify-content: center;
-  align-items: center;
-}
-
-.reservation-list {
-  position: relative;
-  top: -19px;
-  min-height: 280px;
-  max-height: 280px;
-  width: 300px;
-  border: 1px;
-  border-radius: 10px;
-  box-shadow: rgba(0, 0, 0, 0.15) 0 15px 25px, rgba(0, 0, 0, 0.05) 0 5px 10px;
-  overflow: auto;
-}
-
-.reservation-list-title {
-  text-align: center;
-}
-
-.summary {
-  font-size: 11px;
-  font-weight: 100;
-}
-
-@keyframes animitop {
-  from {
-    top: -300px;
-    opacity: 0;
-  }
-
-  to {
-    top: 0;
-    opacity: 1;
-  }
-}
