+++ src/tests/commentsCounter.test.js
