--- src2/index.html
@@ -1,90 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <script src="index.js" type="module" defer></script>
-  <!-- <script src="https://cdn.tailwindcss.com"></script> -->
-  <script src="https://kit.fontawesome.com/9b54de88d2.js" crossorigin="anonymous"></script>
-  <link rel="stylesheet" href="style.css">
-  <title>TV Zone</title>
-</head>
-
-<body>
-  <header>
-    <nav>
-      <button type="button" class="burger-menu" data-toggled="false">
-        <div class="line line1"></div>
-        <div class="line line2"></div>
-        <div class="line line3"></div>
-      </button>
-      <ul class="nav-list hidden">
-        <li><a class="nav-link" href="./index.html">Home</a></li>
-        <li><a class="nav-link" href="#shows-section">Tv Shows</a></li>
-        <li><a class="nav-link" href="#contact">Contact</a></li>
-      </ul>
-    </nav>
-  </header>
-  <main>
-    <div class="headline">
-      <h1>Welcome to TV Zone</h1>
-      <h3>
-        Where you can find all the information about your favourite series
-      </h3>
-      <button type="button">Start browsing</button>
-    </div>
-  </main>
-  <section id="shows-section" class="shows-section">
-    <h2 class="shows-header">Shows</h2>
-    <div class="shows"></div>
-    <div class="shows-controls">
-      <button type="button" class="page-control previous">Previous</button>
-      <span class="page-number">1</span>
-      <button type="button" class="page-control next">Next</button>
-    </div>
-  </section>
-  <footer id="contact">
-    <p>Created by:</p>
-    <ul>
-      <li>
-        <a href="https://github.com/laiifuu">Hayet Fellah</a>
-      </li>
-      <li>
-        <a href="https://github.com/Zuhaib042">Zuhaib Amjad</a>
-      </li>
-      <li>
-        <a href="https://github.com/KareemWilson">Kareem Wilson</a>
-      </li>
-    </ul>
-  </footer>
-  <!-- <div class="popup">
-    <button class="close-btn"></button>
-    <section class="show-summary">
-      <div class="shows-poster">
-        <img src="" alt="show poster">
-      </div>
-      <div class="show-info">
-        <h2  class="title">title</h2>
-        <ul class="genres"></ul>
-        <div>Rating: <span class="rating">rating</span></div>
-        <p class="summary">add summary here</p>
-      </div>
-    </section>
-    <section class="add-comment-section">
-      <h2>Add comment: </h2>
-      <form action="">
-        <input type="text">
-        <textarea name="" id="" cols="30" rows="10"></textarea>
-        <button></button>
-      </form>
-    </section>
-    <section class="comments-section">
-      <h2>Comments: </h2>
-      <ul class="comments-list"></ul>
-    </section>
-  </div> -->
-</body>
-
-</html>
