--- src2/style.css
@@ -1,436 +1,0 @@
-:root {
-  --orange: #ff8906;
-  --black: #0f0e17;
-  --grey: #a7a9be;
-  --pink: #e53170;
-  --form-input: #0f0e17;
-  --paragraph: #2e2f3e;
-  --card-headline: #fffffe;
-  --fs-200: 14px;
-  --fs-300: 16px;
-  --fs-400: 18px;
-  --fs-500: 20px;
-  --fs-600: 24px;
-  --fs-700: 30px;
-  --fs-800: 36px;
-}
-
-*,
-*::before,
-*::after {
-  box-sizing: border-box;
-
-  /* border: solid 1px pink;  */
-}
-
-body,
-h1,
-h2,
-h3,
-h4,
-p,
-figure,
-blockquote,
-dl,
-dd {
-  margin: 0;
-}
-
-ul,
-ol {
-  list-style: none;
-  margin: 0;
-  padding: 0;
-}
-
-html {
-  scroll-behavior: smooth;
-}
-
-body {
-  min-height: 100vh;
-  text-rendering: optimizeSpeed;
-  line-height: 1.5;
-  font-family: "Poppins", sans-serif;
-  color: var(--card-headline);
-  background-color: var(--black);
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-a:not([class]) {
-  color: var(--pink);
-}
-
-img,
-picture {
-  max-width: 100%;
-  display: block;
-}
-
-input,
-button,
-textarea,
-select {
-  font: inherit;
-}
-
-header {
-  position: absolute;
-  top: 20px;
-  left: 20px;
-  right: 20px;
-}
-
-nav {
-  display: flex;
-  align-items: center;
-}
-
-.nav-list {
-  display: none;
-}
-
-.burger-menu {
-  width: 40px;
-  height: 30px;
-  padding: 0;
-  color: var(--card-headline);
-  background-color: transparent;
-}
-
-.line {
-  width: 40px;
-  height: 6px;
-  color: var(--card-headline);
-  background-color: var(--card-headline);
-  transition: all 0.3s ease-in-out;
-}
-
-.line1 {
-  transform: translateY(-6px);
-}
-
-.line3 {
-  transform: translateY(6px);
-}
-
-.nav-link {
-  padding: 15px;
-}
-
-main {
-  height: 100vh;
-  background-image:
-    linear-gradient(rgba(0, 0, 0, 0.3), rgba(0, 0, 0, 0.3)),
-    url("./images/once-upon-a-time.jpg");
-  background-repeat: no-repeat;
-  background-size: cover;
-  padding: 20px;
-}
-
-.headline {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  height: 100vh;
-  gap: 20px;
-}
-
-.shows-section {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 20px;
-  margin-top: 40px;
-}
-
-.shows-header{
-  font-weight: 600px;
-}
-
-.shows {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 40px;
-  justify-content: center;
-  align-items: center;
-  padding: 20px;
-}
-
-.show-card {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  width: 100%;
-  height: auto;
-}
-
-.interactions-section {
-  height: 100%;
-  width: 100%;
-  padding: 5px;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  right: 0;
-  z-index: 2;
-  display: flex;
-  justify-content: space-around;
-  align-items: flex-end;
-  background-image: linear-gradient(transparent, rgba(0, 0, 0, 1));
-}
-
-button {
-  padding: 10px;
-  border: none;
-  background-color: var(--pink);
-  color: var(--card-headline);
-  cursor: pointer;
-}
-
-.like-btn {
-  background-color: transparent;
-  color: var(--pink);
-  width: 48px;
-}
-
-.likes-number {
-  font-weight: 700;
-  color: var(--pink);
-}
-
-.info-btns {
-  display: flex;
-  gap: 10px;
-}
-
-
-.page-control {
-  background-color: transparent;
-  color: var(--grey);
-  text-decoration: underline;
-  font-weight: 700;
-}
-
-footer {
-  margin-top: 50px;
-  display: flex;
-  flex-wrap: wrap;
-  gap: 20px;
-  padding: 30px;
-}
-
-footer > ul {
-  display: flex;
-  gap: 10px;
-}
-
-.hidden {
-  display: none;
-}
-
-/* popupstyling */
-.popup{
-  position: absolute;
-  top: 0;
-  z-index: 5;
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-  padding: 20px;
-  justify-content: center;
-  align-items: center;
-  background-color: var(--black);
-  overflow: scroll;
-}
-
-.no-scroll {
-  overflow: hidden;
-}
-
-.close-btn {
-  position: absolute;
-  top: 0;
-  right: 0;
-  background-color: transparent;
-}
-
-.show-summary{
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.show-info{
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 10px;
-}
-
-.genres {
-  display: flex;
-  gap: 5px;
-}
-
-.genre-item{
-  background-color: var(--pink);
-  padding: 5px;
-  border-radius: 20px;
-  border: 1px solid var(--card-headline);
-  font-size: var(--fs-200);
-  font-weight: 600;
-}
-
-.add-comment-section, .add-reservation{
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-  width: 100%;
-}
-
-.popup-form{
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: stretch;
-  gap:10px;
-}
-
-input, textarea{
-  padding: 10px;
-  background-color: #a7a9be;
-  border: none;
-}
-
-.comments-section, .reservations-section{
-  width: 100%;
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.comments-list{
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-.comments-list > li, .reservation-list>li{
-  padding: 10px;
-  border-bottom: 1px solid var(--grey);
-}
-
-.commment-date, .user-name{
-  font-weight: 600;
-  color: var(--pink);
-}
-
-.date{
-  color: var(--grey);
-  font-weight: 600;
-}
-
-@media screen and (min-width: 600px) {
-  .shows {
-    align-items: stretch;
-  }
-
-  .show-card{
-    width: 40%;
-  }
-
-  .info-btns{
-    flex-direction: column;
-  }
-}
-
-@media screen and (min-width: 768px) {
-  h1 {
-    font-size: 48px;
-  }
-
-  h2 {
-    font-size: 36px;
-  }
-
-  nav {
-    justify-content: flex-end;
-  }
-
-  .burger-menu {
-    display: none;
-  }
-
-  .nav-list {
-    display: flex;
-    gap: 20px;
-    font-weight: 600;
-    justify-content: flex-end;
-  }
-
-  .nav-link {
-    text-decoration: none;
-    color: var(--card-headline);
-  }
-
-  main {
-    padding-left: 100px;
-  }
-
-  .shows {
-    gap: 40px;
-    align-items: stretch;
-  }
-
-  .show-card {
-    width: 25%;
-  }
-
-  .img-placeholder > img {
-    height: auto;
-  }
-
-  .info-buttons {
-    flex-direction: column;
-  }
-
-  .info-btns > button {
-    padding: 5px;
-  }
-
-  .popup{
-    flex-wrap: wrap;
-    flex-direction: row;
-    gap: 20px;
-    align-items: flex-start;
-  }
-  .show-summary {
-    flex-direction: row;
-  }
-
-  .shows-poster{
-    flex: 1;
-  }
-
-  .show-info{
-    flex: 2;
-    align-items: flex-start;
-  }
-
-  .add-comment-section, .add-reservation{
-    flex: 0.5;
-  }
-
-  .comments-section, .reservations-section{
-    flex: 0.5;
-  }
-}
